pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobook-cli"
version = "0.1.0"
dependencies = ["pyyaml", "ffmpeg-python"]

[project.scripts]
audiobook-tool = "audiobook.app:main"
